Avatar and Player Lv added to the BoardPvP Player Info Panel
The BoardPvP UI has been updated to show the player's avatar and Lv while still showing the Symbol image.
LP has been updated to be slighly bigger and colored yellow for better visibility

+++ DungeonDiceMonsters/Data/GameData.cs
@@ -197,7 +197,7 @@ namespace DungeonDiceMonsters
         }
         public static string GetPlayerDataForPvPMatch()
         {
-            return string.Format("{0}|{1}|{2}", _PlayerName, _Playerlevel, (int)_PlayerAvatar);
+            return string.Format("{0}|{1}|{2}", _PlayerName, _Playerlevel, _PlayerAvatar);
         }
         #endregion
